Moved test to correct file
